CFY-4987 adapt tests to ctx throwing an error on undefined runtime props
using undefined shell variables in a "ctx" invocation makes the call be
a "get" rather than a "set".

Ensure a "set" operation by defaulting the shell variable to an empty
string, making the whole command always have 2 arguments.

This change also does "set -u" for the whole test script, in hope to
prevent similar errors in the future.