CY-2457 Update fabric to 2.x, refactor (#67)
* Update dependencies

* Rewrite run_script

* Drop tunnel.py

* update run_task & run_module_task

* drop unused hide_output

* update run_commands

* drop unused funcs

* use 2+3-compat exec()

* Bump version to 2.0.0

* flake

* output logging

* Fix unittests

* drop FabricPluginRealSSHTests

they haven't been running anyway. And i'm not gonna fix them now.

* unittests: drop test_hide_* tests

hide isnt a feature anymore

* unittests: drop test_failed_command

it was asserting that an exception was thrown, which was thrown by
a mock. No need to test mocks.

* drop envvar for key path

wtf with that feature

* drop test_warn_only

warn_only is not a feature anymore

* change ssh connection creating logic

fabric env defaults, figuring out fabric_env, etc

* update getting fabric tasks

* add docstrings

* CI: remove unused file

* update dev-reqs

* CI: drop tox

* drop useless entries from the example

* drop `hide_output` from the blueprint